Improvements, new features, what's next?
This commit is quite a jump from the previous commit, as I forgot to push changes once I had created the "about" form. I will still be pushing this as v0.3, and the reason for that will be reflected by the notes below.

Total changed files: 13.

Additionals:
+ Introduced Form2.cs as [2]QuickFix_about.cs.
+ Added text, button and hyperlink to [2]QuickFix_about.cs.
+ Introduced [99]QuickFix_testing.cs where functionality is to be tested without damaging the rest of the project.

Changes:
~ Renamed Form1.cs to [1]QuickFix_main.cs.
~ A [99]QuickFix_testing.cs. can be found, but is not implemented.
~ Replaced btnDebugYes functionality in favour of access to [99]QuickFix_testing.cs.

This is not the extensive list of changes, however due to the renaming of certain things, I do not think git is going to mark the changes made (as it will register as Form1 being deleted and [1]QuickFix_main being created). This is a limitation I can't beat, so as a work around, I will start including a changelog.md with the next push.

A new branch will be opening soon, and eventually this branch will become `staging-alpha`. This new branch will push with the updated UI. At that point, all references to this branch will be destroyed but will remain for educational reasons.

And here it begins to take shape!