Beta is open!
This update is completely re-written from the ground up, new layout, more intuitive.

This will go down as release 0.5.0. Audio and Networking should be working fine.

Please note that the reboot buttons only echo their command right now. This is only for testing reasons, nobody wants to reboot their PC just to test if a button works, when we know that the command works.

Now we move on to solving Issue #1 and adding other functionality.

Hazzah!